backend: created added /user/* to untrusted policies
The RBAC.CreateMultiple function now allows multiple subjects to be
provided, and a number of RBAC policies will be created to meet the
definition. This is similar to the grouping create multiple function.

This allows the /user/* authorization subject to be added to all
policies which /untrusted_user was a subject. Giving logged in users
at the same permissions as untrusted users.