backend: made authorization init ignore user groupings
The groupings reconcile login in the AuthorizationClient was including
user groupings. This resulted in the logic deleting user <> role
associations because they weren't declared in the GROUPING_POLICIES
variable.

This change makes the Reconcile listAll function not return any
groupings where the subject is a user. Since those should be the only
types of groupings not declared in the previously mentioned variable.